% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_under_curve.R
\name{area_under_curve}
\alias{area_under_curve}
\title{Area Under The Curve}
\usage{
area_under_curve(fpr, tpr, ratio_name, plot.ROC)
}
\arguments{
\item{fpr}{vector with values of the false positive rate for all possible thresholds (or a range of thresholds)}

\item{tpr}{vector with values of the true positive rate for all possible thresholds (or a range of thresholds)}

\item{ratio_name}{name of the band ratio}

\item{plot.ROC}{logical. If true, the ROC is plotted}
}
\description{
calculates the area under the curve and plots the result. Based on the results of \emph{sensitivity_analysis}
}
\examples{
  auc <- area_under_curve(fpr = sensitivity_analysis[,5], tpr = sensitivity_analysis[,4],  plot.ROC = TRUE)
}
