% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_glacier_area.R
\name{calc_glacier_area}
\alias{calc_glacier_area}
\title{Calculate the glacierized area}
\usage{
calc_glacier_area(x, res_x, res_y)
}
\arguments{
\item{x}{raster with glacierized pixels containing the value 1}

\item{res_x}{resolution of the raster in x direction in km. By default 0.03 (resolution of Landsat images).}

\item{res_y}{resolution of the raster in y direction in km. By default 0.03 (resolution of Landsat images).}
}
\description{
Calculates the glacierized area (in km²) from a map where the glacier is indicated by the value 1
}
\examples{
glacier_area <- calc_glacier_area(binary_map_ndsi)
}
