% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ratio.R
\name{calc_ratio}
\alias{calc_ratio}
\title{Calcuate a band ratio}
\usage{
calc_ratio(raster, ratio, bnbr_green, bnbr_red, bnbr_nir, bnbr_swir, bnbr_tir)
}
\arguments{
\item{raster}{raster of which the band ratio should be calculated of}

\item{ratio}{name of the band ratio. Possible values: "red/swir", "nir/swir", "tir/(nir/swir)" and "ndsi".}

\item{bnbr_green}{number of the green channel (obligatory for "ndsi" ratio)}

\item{bnbr_red}{number of the green channel (obligatory for "red/swir" ratio)}

\item{bnbr_nir}{number of the green channel (obligatory for "nir/swir" and "tir/(nir/swir)" ratios)}

\item{bnbr_swir}{number of the green channel (obligatory for all ratios)}

\item{bnbr_tir}{number of the green channel (obligatory for "tir/(nir/swir)" ratio)}
}
\description{
Caculates band ratios of a raster which are known to be sensitive to glacierized areas.
}
\examples{
band_ratio <- calc_ratio(raster = landsat_images, ratio = "red/swir", bnbr_red = 4, bnbr_swir = 6)
}
