% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findthreshold.R
\name{find_threshold}
\alias{find_threshold}
\title{Find threshold for a band ratio}
\usage{
find_threshold(raster, plot.hist)
}
\arguments{
\item{raster}{raster (ideally with a bimodel distribution of values)}

\item{plot.hist}{logical. If true, histogram with threshold indicated in darkred is plotted. For visual check of the validity of the threshold.}
}
\description{
{Finds the as best considered threshold based on the histogram of the raster.
The histogram is considered to be bimodal representing the two classes.
The best threshold is considered to be at position of the local minimum between the two humps.}
}
