% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_binary_map.R
\name{get_binary_map}
\alias{get_binary_map}
\title{Derive a binary map}
\usage{
get_binary_map(ratio, threshold, ratio_name)
}
\arguments{
\item{ratio}{raster of the band ratio}

\item{threshold}{threshold that should be applied on the raster}

\item{ratio_name}{name of the ratio.  Possible values: "red/swir", "nir/swir", "tir/(nir/swir)" and "ndsi".}
}
\description{
Get a binary map (glacier/no glacier) from a raster containing the values of a band ratio. Based on a given threshold.
}
\examples{
binary_map <- get_binary_map(ratio = ndsi_ratio, threshold = 0.4, ratio_name = "ndsi")
}
