% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_min_and_max_threshold.R
\name{get_min_and_max_threshold}
\alias{get_min_and_max_threshold}
\title{Get the minimum and maximum threshold based on a criterion}
\usage{
get_min_and_max_threshold(thresholds, values, max_diff, value, min_value)
}
\arguments{
\item{thresholds}{vector of all possible thresholds}

\item{values}{vector of values which should meet the criterion}

\item{max_diff}{maximum difference from a value in percent (scenario a)}

\item{value}{value from which the column value can derogate to a certain degree (scenario a)}

\item{min_value}{the value tha values vector should not fall below (scenario b)}
}
\description{
Based on the results of \emph{\code{\link[=sensitivity_analysis]{sensitivity analysis}}}. Derives the minimum and maximum
threshold meeting one of the following criteria: a) maximum difference of one column from a value or b)
minimum value of a column.
}
