% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadandinstall.R
\name{loadandinstall}
\alias{loadandinstall}
\title{Load and install packages}
\usage{
loadandinstall(pkglist)
}
\arguments{
\item{pkglist}{vector with all needed packages}
}
\description{
Loads all packages in a list and installs them first in case they are not installed yet.
}
\examples{
packages <- c("ggplot2", "raster")
loadandinstall(packages)
}
