% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{sensitivity_analysis}
\alias{sensitivity_analysis}
\title{Sensitivity analysis of thresholding a band ratio}
\usage{
sensitivity_analysis(raster, val_raster, ratio_name, threshold_min,
  threshold_max, threshold, steps)
}
\arguments{
\item{raster}{raster of the band ratio}

\item{val_raster}{if available: raster with the ground truth as validation data}

\item{ratio_name}{name of the ratio (how \emph{raster} is calculated)}

\item{threshold_min}{minimum threshold. Lower end of the threshold range.}

\item{threshold_max}{maximum threshold. Upper end of the threshold range.}

\item{steps}{number of steps within the range of thresholds}
}
\description{
Calculates the glacier area for a range of thresholds. If validation data is available,
it also caculates accuracy measurements (overall accuracy, false positive ratio and true positive ratio).
The results for the FPR and TPR can then be used to calculate the area under the curve (AOC).
}
