% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftvector.R
\name{shift.vec}
\alias{shift.vec}
\title{Shift a vector by a given number of positions}
\usage{
shift.vec(vec, shift)
}
\arguments{
\item{vec}{vector that is to be shifted}

\item{shift}{number of positions that the vector should be shifted. Positive and negative values possible. New position of the value: original position + \emph{shift}}
}
\description{
Shifts a vector by a given number of positions. Unfilled positions are replaced by \emph{NA}.
}
\examples{
#get the previous value
previous <- shift.vec(vector, 1)
#get the next value
next <- shift.vec(vector, -1)
}
